IOS-7246 Add `EthereumNetworkProvider+`
